requirements.txt
Pillow==9.0.0
imagehash==4.2.1
termcolor==1.1.0
